An experimental implementation of factor expressions for table and vector/matrix factors.
This revision introduces major changes to the factor framework. The factors are evaluated lazily, so that expressions such as f * g * 2 are evaluated in a single pass, rather than creating a temporary for f * g. This provides substantial speed-ups and allows us to reuse memory when reallocation is not necessary.

With this change, libgm now relies heavily on C++14 features. The support for older compilers, including GCC 4.8, has been discontinued.